requirements.txt
boto3==1.34.0
Flask==2.3.3
RPi.GPIO==0.7.1
requests==2.31.0
python-dotenv==1.0.0